lvis-engine/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
Pillow>=10.3.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
scikit-image>=0.23.0
PyExifTool>=0.5.6
# RAW camera format support
rawpy>=0.22.0
imageio>=2.34.0
